Update README.md
This project involves building a deep learning model to classify clothing images using the Fashion MNIST dataset. The dataset is a more challenging and realistic alternative to the traditional MNIST digit dataset, consisting of 70,000 grayscale images of 10 different clothing categories such as T-shirts, trousers, sneakers, and bags. Each image is 28x28 pixels in size, and the dataset is split into 60,000 training and 10,000 testing samples.